Moved type declarations out of state and into types files
